初步完成修复

+++ src/com/sz/plugin/manager/ArtifactManager.java
@@ -61,7 +61,7 @@ public class ArtifactManager {
     private long getDamageCap() throws Exception {
         long lm;
         try {
-            Object item = MSUtils.doMethod(this.player, "getInventorySlot", (byte)-1, -11);
+            Object item = MSUtils.doMethod(this.player, "getInventorySlot", (byte)-1, (short)-11);
             lm = ((Long)MSUtils.doMethod(item, "getLimitBreak")).longValue();
         }
         catch (Exception e)
